+++ app/models/article.rb
@@ -24,16 +24,4 @@ class Article < ApplicationRecord
   belongs_to :user
   has_one_attached :eyecatch
   has_rich_text :content
-
-  def display_created_at
-    I18n.l(self.created_at, format: :default)
-  end
-
-  def author_name
-    user.display_name
-  end
-
-  def like_count
-    likes.count
-  end
 end
