+++ app/models/user.rb
@@ -36,25 +36,12 @@ class User < ApplicationRecord
 
   delegate :birthday, :age, :gender, to: :profile, allow_nil: true
 
-  # :articlesは複数形
-
   def has_written?(article)
     articles.exists?(id: article.id)
   end
 
   def has_liked?(article)
     likes.exists?(article_id: article.id)
-  end
-
-  def display_name
-    # if profile && profile.nickname
-    #   profile.nickname
-    # else
-    #   self.email.split('@').first
-    # end
-
-    # ぼっち演算子
-    profile&.nickname || self.email.split('@').first
   end
 
   def follow!(user)
@@ -72,24 +59,8 @@ class User < ApplicationRecord
     following_relationships.exists?(following_id: user.id)
   end
 
-  # def birthday
-  #   profile&.birthday
-  # end
-
-  # def gender
-  #   profile&.gender
-  # end
-
   def prepare_profile
     profile || build_profile
-  end
-
-  def avatar_image
-    if profile&.avatar&.attached?
-      profile.avatar
-    else
-      'default-avatar.png'
-    end
   end
 
   private
